+++ ext/parquet/src/types/record_types.rs
@@ -1,5 +1,7 @@
 use itertools::Itertools;
 
+use crate::reader::ReaderError;
+
 use super::*;
 
 #[derive(Debug)]
@@ -15,15 +17,16 @@ pub enum ColumnRecord<S: BuildHasher + Default> {
 }
 
 #[derive(Debug)]
-pub struct ParquetField(pub Field);
+pub struct ParquetField(pub Field, pub bool);
 
-impl<S: BuildHasher + Default> IntoValue for RowRecord<S> {
-    fn into_value_with(self, handle: &Ruby) -> Value {
+impl<S: BuildHasher + Default> TryIntoValue for RowRecord<S> {
+    fn try_into_value_with(self, handle: &Ruby) -> Result<Value, ReaderError> {
         match self {
             RowRecord::Vec(vec) => {
                 let ary = handle.ary_new_capa(vec.len());
-                vec.into_iter().try_for_each(|v| ary.push(v)).unwrap();
-                handle.into_value(ary)
+                vec.into_iter()
+                    .try_for_each(|v| ary.push(v.try_into_value_with(handle)?))?;
+                Ok(handle.into_value(ary))
             }
             RowRecord::Map(map) => {
                 let hash = handle.hash_new_capa(map.len());
@@ -36,41 +39,41 @@ impl<S: BuildHasher + Default> IntoValue for RowRecord<S> {
                     for (k, v) in chunk {
                         if i + 1 >= values.len() {
                             // Bulk insert current batch if array is full
-                            hash.bulk_insert(&values[..i]).unwrap();
+                            hash.bulk_insert(&values[..i])?;
                             values[..i].fill(handle.qnil().as_value());
                             i = 0;
                         }
                         values[i] = handle.into_value(k);
-                        values[i + 1] = handle.into_value(v);
+                        values[i + 1] = v.try_into_value_with(handle)?;
                         i += 2;
                     }
                     // Insert any remaining pairs
                     if i > 0 {
-                        hash.bulk_insert(&values[..i]).unwrap();
+                        hash.bulk_insert(&values[..i])?;
                         values[..i].fill(handle.qnil().as_value());
                         i = 0;
                     }
                 }
 
-                hash.into_value_with(handle)
+                Ok(hash.into_value_with(handle))
             }
         }
     }
 }
 
-impl<S: BuildHasher + Default> IntoValue for ColumnRecord<S> {
-    fn into_value_with(self, handle: &Ruby) -> Value {
+impl<S: BuildHasher + Default> TryIntoValue for ColumnRecord<S> {
+    fn try_into_value_with(self, handle: &Ruby) -> Result<Value, ReaderError> {
         match self {
             ColumnRecord::Vec(vec) => {
                 let ary = handle.ary_new_capa(vec.len());
-                vec.into_iter()
-                    .try_for_each(|v| {
-                        let nested_ary = handle.ary_new_capa(v.len());
-                        v.into_iter().try_for_each(|v| nested_ary.push(v)).unwrap();
-                        ary.push(nested_ary.into_value_with(handle))
-                    })
-                    .unwrap();
-                ary.into_value_with(handle)
+                vec.into_iter().try_for_each(|v| {
+                    let nested_ary = handle.ary_new_capa(v.len());
+                    v.into_iter()
+                        .try_for_each(|v| nested_ary.push(v.try_into_value_with(handle)?))?;
+                    ary.push(nested_ary.into_value_with(handle))?;
+                    Ok::<_, ReaderError>(())
+                })?;
+                Ok(ary.into_value_with(handle))
             }
             ColumnRecord::Map(map) => {
                 let hash = handle.hash_new_capa(map.len());
@@ -83,91 +86,98 @@ impl<S: BuildHasher + Default> IntoValue for ColumnRecord<S> {
                     for (k, v) in chunk {
                         if i + 1 >= values.len() {
                             // Bulk insert current batch if array is full
-                            hash.bulk_insert(&values[..i]).unwrap();
+                            hash.bulk_insert(&values[..i])?;
                             values[..i].fill(handle.qnil().as_value());
                             i = 0;
                         }
                         values[i] = handle.into_value(k);
                         let ary = handle.ary_new_capa(v.len());
-                        v.into_iter().try_for_each(|v| ary.push(v)).unwrap();
+                        v.into_iter()
+                            .try_for_each(|v| ary.push(v.try_into_value_with(handle)?))?;
                         values[i + 1] = handle.into_value(ary);
                         i += 2;
                     }
                     // Insert any remaining pairs
                     if i > 0 {
-                        hash.bulk_insert(&values[..i]).unwrap();
+                        hash.bulk_insert(&values[..i])?;
                         values[..i].fill(handle.qnil().as_value());
                         i = 0;
                     }
                 }
 
-                hash.into_value_with(handle)
+                Ok(hash.into_value_with(handle))
             }
         }
     }
 }
 
-impl IntoValue for ParquetField {
-    fn into_value_with(self, handle: &Ruby) -> Value {
+pub trait TryIntoValue {
+    fn try_into_value_with(self, handle: &Ruby) -> Result<Value, ReaderError>;
+}
+
+impl TryIntoValue for ParquetField {
+    fn try_into_value_with(self, handle: &Ruby) -> Result<Value, ReaderError> {
         match self.0 {
-            Field::Null => handle.qnil().as_value(),
-            Field::Bool(b) => b.into_value_with(handle),
-            Field::Short(s) => s.into_value_with(handle),
-            Field::Int(i) => i.into_value_with(handle),
-            Field::Long(l) => l.into_value_with(handle),
-            Field::UByte(ub) => ub.into_value_with(handle),
-            Field::UShort(us) => us.into_value_with(handle),
-            Field::UInt(ui) => ui.into_value_with(handle),
-            Field::ULong(ul) => ul.into_value_with(handle),
-            Field::Float16(f) => f32::from(f).into_value_with(handle),
-            Field::Float(f) => f.into_value_with(handle),
-            Field::Double(d) => d.into_value_with(handle),
-            Field::Str(s) => s.into_value_with(handle),
-            Field::Byte(b) => b.into_value_with(handle),
-            Field::Bytes(b) => handle.str_from_slice(b.data()).as_value(),
+            Field::Null => Ok(handle.qnil().as_value()),
+            Field::Bool(b) => Ok(b.into_value_with(handle)),
+            Field::Short(s) => Ok(s.into_value_with(handle)),
+            Field::Int(i) => Ok(i.into_value_with(handle)),
+            Field::Long(l) => Ok(l.into_value_with(handle)),
+            Field::UByte(ub) => Ok(ub.into_value_with(handle)),
+            Field::UShort(us) => Ok(us.into_value_with(handle)),
+            Field::UInt(ui) => Ok(ui.into_value_with(handle)),
+            Field::ULong(ul) => Ok(ul.into_value_with(handle)),
+            Field::Float16(f) => Ok(f32::from(f).into_value_with(handle)),
+            Field::Float(f) => Ok(f.into_value_with(handle)),
+            Field::Double(d) => Ok(d.into_value_with(handle)),
+            Field::Str(s) => {
+                if self.1 {
+                    Ok(simdutf8::basic::from_utf8(s.as_bytes())
+                        .map_err(|e| ReaderError::Utf8Error(e))
+                        .and_then(|s| Ok(s.into_value_with(handle)))?)
+                } else {
+                    Ok(handle.str_from_slice(s.as_bytes()).as_value())
+                }
+            }
+            Field::Byte(b) => Ok(b.into_value_with(handle)),
+            Field::Bytes(b) => Ok(handle.str_from_slice(b.data()).as_value()),
             Field::Date(d) => {
-                let ts = jiff::Timestamp::from_second((d as i64) * 86400).unwrap();
+                let ts = jiff::Timestamp::from_second((d as i64) * 86400)?;
                 let formatted = ts.strftime("%Y-%m-%d").to_string();
-                formatted.into_value_with(handle)
+                Ok(formatted.into_value_with(handle))
             }
             Field::TimestampMillis(ts) => {
-                let ts = jiff::Timestamp::from_millisecond(ts).unwrap();
+                let ts = jiff::Timestamp::from_millisecond(ts)?;
                 let time_class = handle.class_time();
-                time_class
-                    .funcall::<_, _, Value>("parse", (ts.to_string(),))
-                    .unwrap()
-                    .into_value_with(handle)
+                Ok(time_class
+                    .funcall::<_, _, Value>("parse", (ts.to_string(),))?
+                    .into_value_with(handle))
             }
             Field::TimestampMicros(ts) => {
-                let ts = jiff::Timestamp::from_microsecond(ts).unwrap();
+                let ts = jiff::Timestamp::from_microsecond(ts)?;
                 let time_class = handle.class_time();
-                time_class
-                    .funcall::<_, _, Value>("parse", (ts.to_string(),))
-                    .unwrap()
-                    .into_value_with(handle)
+                Ok(time_class
+                    .funcall::<_, _, Value>("parse", (ts.to_string(),))?
+                    .into_value_with(handle))
             }
             Field::ListInternal(list) => {
                 let elements = list.elements();
                 let ary = handle.ary_new_capa(elements.len());
-                elements
-                    .iter()
-                    .try_for_each(|e| ary.push(ParquetField(e.clone()).into_value_with(handle)))
-                    .unwrap();
-                ary.into_value_with(handle)
+                elements.iter().try_for_each(|e| {
+                    ary.push(ParquetField(e.clone(), self.1).try_into_value_with(handle)?)
+                })?;
+                Ok(ary.into_value_with(handle))
             }
             Field::MapInternal(map) => {
                 let entries = map.entries();
                 let hash = handle.hash_new_capa(entries.len());
-                entries
-                    .iter()
-                    .try_for_each(|(k, v)| {
-                        hash.aset(
-                            ParquetField(k.clone()).into_value_with(handle),
-                            ParquetField(v.clone()).into_value_with(handle),
-                        )
-                    })
-                    .unwrap();
-                hash.into_value_with(handle)
+                entries.iter().try_for_each(|(k, v)| {
+                    hash.aset(
+                        ParquetField(k.clone(), self.1).try_into_value_with(handle)?,
+                        ParquetField(v.clone(), self.1).try_into_value_with(handle)?,
+                    )
+                })?;
+                Ok(hash.into_value_with(handle))
             }
             Field::Decimal(d) => {
                 let value = match d {
@@ -185,20 +195,24 @@ impl IntoValue for ParquetField {
                         format!("{}e-{}", unscaled, scale)
                     }
                 };
-                handle.eval(&format!("BigDecimal(\"{value}\")")).unwrap()
+                Ok(handle.eval(&format!("BigDecimal(\"{value}\")"))?)
             }
             Field::Group(row) => {
                 let hash = handle.hash_new();
-                row.get_column_iter()
-                    .try_for_each(|(k, v)| {
-                        hash.aset(
-                            k.clone().into_value_with(handle),
-                            ParquetField(v.clone()).into_value_with(handle),
-                        )
-                    })
-                    .unwrap();
-                hash.into_value_with(handle)
+                row.get_column_iter().try_for_each(|(k, v)| {
+                    hash.aset(
+                        k.clone().into_value_with(handle),
+                        ParquetField(v.clone(), self.1).try_into_value_with(handle)?,
+                    )
+                })?;
+                Ok(hash.into_value_with(handle))
             }
         }
     }
 }
+
+// impl IntoValue for ParquetField {
+//     fn into_value_with(self, handle: &Ruby) -> Value {
+//         self.try_into_value_with(handle).unwrap()
+//     }
+// }